+++ analysis/analyzer_test.go
@@ -10,10 +10,10 @@ import (
 
 type MockElevationQuerier struct{}
 
-func (m *MockElevationQuerier) QueryElevations(_ context.Context, points orb.LineString) ([]int32, error) {
-	var out []int32
+func (m *MockElevationQuerier) QueryElevations(_ context.Context, points orb.LineString) ([]float64, error) {
+	var out []float64
 	for range points {
-		out = append(out, 42)
+		out = append(out, 42.0)
 	}
 	return out, nil
 }
@@ -28,5 +28,5 @@ func TestAnalyzer_HydrateTrack(t *testing.T) {
 
 	props := got.Properties
 	require.Equal(t, 157425.537108, props["lengthMeters"])
-	require.Equal(t, []int32{42, 42}, props.CoordinateProperties()["elevationMeters"])
+	require.Equal(t, []float64{42.0, 42.0}, props.CoordinateProperties()["elevationMeters"])
 }
